+++ src/main.tsx
@@ -2,7 +2,7 @@ import '@/styles/index.css';
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { ErrorBoundary } from '@/components/ErrorBoundary';
-import { AppRouter } from '@/router';
+import { AppRouter } from '@/router/AppRouter';
 
 const root = document.getElementById('root');
 if (!root) throw new Error('Root element not found');
