+++ src/router/appRoutes.tsx
@@ -1,0 +1,4 @@
+export const AppRoutes = {
+  HOME: '/',
+  DETAILS: '/details',
+} as const;
