--- src/router/index.ts
@@ -1,3 +1,0 @@
-export { AppRouter } from './AppRouter';
-export { routes } from './routes';
-export { MainLayout } from './MainLayout';
